Initial implementation of event dispatcher.
Dispatcher implementation in Dispatch\Dispatcher. The dispatcher allows
registering implementations of Dispatch\Task to unique task identifiers,
for a class of event. When an event is dispatched by the dispatcher, it
will execute all (or specific) registered tasks for that event
asynchronously.

A Dispatch\Task is built up of sub-tasks, which implement the EventListening
interface. There are two implementations for a Task: AsynchronousTask
and SynchronousTask. An AsynchronousTask will propagate the event to all
of it's subtasks asynchronously, stopping propagation is not allowed as
it is unsafe. A SynchronousTask will enforce that subtasks handle the
event synchronously; and therefore listeners can be added with priority
over one another, allowing higher priority listeners to stop propagation
of the event for the rest of the task.